Move def_id_to_path to use site in visit_ast
